IN-38-copy: refactor: clean code in all files
