+++ app/auth/sing-up/page.tsx
@@ -1,4 +1,4 @@
-import SignUpPage from '@/src/features/auth/sign-up/ui/SignUpPage';
+import SignUpPage from '@/src/features/auth/ui/sign-up/ui/SignUpPage';
 
 export default function Page() {
   return <SignUpPage />;
