+++ src/features/auth/ui/sign-up/ui/SignInPrompt/SignInPrompt.tsx
