+++ src/features/auth/ui/sign-up/ui/SignUpForm/SignUpForm.module.scss
