+++ src/features/auth/ui/sign-up/ui/SignUpPage.module.scss
