+++ src/features/auth/ui/sign-up/ui/SignUpPage.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import { SignUpForm } from '@/src/features/auth/sign-up/ui/SignUpForm/SignUpForm';
+import { SignUpForm } from '@/src/features/auth/ui/sign-up/ui/SignUpForm/SignUpForm';
 import Card from '@/src/shared/ui/Card/Card';
 
 import styles from './SignUpPage.module.scss';
-import { OAuthButtons } from '@/src/features/auth/sign-up/ui/OAuthButtons/OAuthButtons';
-import { SignInPrompt } from '@/src/features/auth/sign-up/ui/SignInPrompt/SignInPrompt';
+import { OAuthButtons } from '@/src/features/auth/ui/sign-up/ui/OAuthButtons/OAuthButtons';
+import { SignInPrompt } from '@/src/features/auth/ui/sign-up/ui/SignInPrompt/SignInPrompt';
 
 export default function SignUpPage() {
   return (
